[wip] derive: move derive expansion to Tree
