+++ miniconf_derive/src/field.rs
@@ -1,8 +1,4 @@
-use darling::{
-    ast::{self, Data},
-    util::{Flag, SpannedValue},
-    Error, FromDeriveInput, FromField, FromVariant,
-};
+use darling::{util::Flag, FromField};
 use proc_macro2::TokenStream;
 use quote::quote;
 
@@ -13,6 +9,8 @@ pub struct TreeField {
     // pub vis: syn::Visibility,
     pub ty: syn::Type,
     // attrs: Vec<syn::Attribute>,
+    #[darling(skip)]
+    pub variant: bool,
     #[darling(default)]
     pub depth: usize,
     pub skip: Flag,
@@ -74,7 +72,13 @@ impl TreeField {
             Some(get) => quote! {
                 #get(self).map_err(|msg| ::miniconf::Traversal::Access(0, msg).into())
             },
-            None => quote! { Ok(&self.#ident) },
+            None => {
+                if self.variant {
+                    quote! { Ok(#ident) }
+                } else {
+                    quote! { Ok(&self.#ident) }
+                }
+            }
         }
     }
 
@@ -84,7 +88,13 @@ impl TreeField {
             Some(get_mut) => quote!(
                 #get_mut(self).map_err(|msg| ::miniconf::Traversal::Access(0, msg).into())
             ),
-            None => quote!( Ok(&mut self.#ident) ),
+            None => {
+                if self.variant {
+                    quote! { Ok(#ident) }
+                } else {
+                    quote! { Ok(&mut self.#ident) }
+                }
+            }
         }
     }
 
@@ -180,170 +190,4 @@ impl TreeField {
             }
         }
     }
-}
-
-#[derive(Debug, FromVariant, Clone)]
-#[darling(attributes(tree))]
-pub struct TreeVariant {
-    pub ident: syn::Ident,
-    pub rename: Option<syn::Ident>,
-    pub skip: Flag,
-    pub fields: ast::Fields<TreeField>,
-}
-
-#[derive(Debug, FromDeriveInput, Clone)]
-#[darling(attributes(tree))]
-#[darling(supports(any))]
-pub struct Tree {
-    pub ident: syn::Ident,
-    pub generics: syn::Generics,
-    // pub vis: syn::Visibility,
-    pub data: ast::Data<TreeVariant, TreeField>,
-    // attrs: Vec<syn::Attribute>,
-    pub tag: Option<SpannedValue<syn::Path>>, // FIXME: implement
-}
-
-impl Tree {
-    pub(crate) fn depth(&self) -> usize {
-        match &self.data {
-            Data::Struct(fields) => depth(&fields.fields) + 1,
-            Data::Enum(variants) => depth(variants.iter().flat_map(|v| v.fields.fields.iter())) + 2,
-        }
-    }
-
-    pub(crate) fn parse(input: &syn::DeriveInput) -> Result<Self, Error> {
-        let mut tree = Self::from_derive_input(input)?;
-
-        match &mut tree.data {
-            Data::Struct(fields) => {
-                if let Some(tag) = &tree.tag {
-                    return Err(Error::custom("No `tag` for structs").with_span(&tag.span()));
-                }
-                remove_skipped(&mut fields.fields)?;
-            }
-            Data::Enum(variants) => {
-                if let Some(tag) = &tree.tag {
-                    return Err(
-                        Error::custom("`tag` for enums unimplemented").with_span(&tag.span())
-                    );
-                }
-                variants.retain(|v| !v.skip.is_present());
-                for v in variants.iter_mut() {
-                    remove_skipped(&mut v.fields.fields)?;
-                }
-            }
-        }
-        Ok(tree)
-    }
-
-    pub(crate) fn fields(&self) -> &Vec<TreeField> {
-        let Data::Struct(fields) = &self.data else {
-            unreachable!()
-        };
-        &fields.fields
-    }
-
-    pub(crate) fn bound_generics<F>(&mut self, func: &mut F)
-    where
-        F: FnMut(usize) -> Option<syn::TypeParamBound>,
-    {
-        match &self.data {
-            Data::Struct(fields) => fields
-                .fields
-                .iter()
-                .for_each(|f| walk_type_params(f.typ(), func, f.depth, &mut self.generics)),
-            Data::Enum(variants) => variants
-                .iter()
-                .flat_map(|v| v.fields.fields.iter())
-                .for_each(|f| walk_type_params(f.typ(), func, f.depth, &mut self.generics)),
-        }
-    }
-}
-
-fn depth<'a>(fields: impl IntoIterator<Item = &'a TreeField>) -> usize {
-    fields.into_iter().fold(0, |d, field| d.max(field.depth))
-}
-
-fn remove_skipped(fields: &mut Vec<TreeField>) -> Result<(), Error> {
-    // unnamed fields can only be skipped if they are terminal
-    while fields
-        .last()
-        .map(|f| f.ident.is_none() && f.skip.is_present())
-        .unwrap_or_default()
-    {
-        fields.pop();
-    }
-    fields.retain(|f| f.ident.is_some() && !f.skip.is_present());
-    if let Some(f) = fields.iter().filter(|f| f.skip.is_present()).next() {
-        Err(
-            Error::custom("Can not `skip` non-terminal tuple struct fields")
-                .with_span(&f.skip.span()),
-        )
-    } else {
-        Ok(())
-    }
-}
-
-fn walk_type_params<F>(typ: &syn::Type, func: &mut F, depth: usize, generics: &mut syn::Generics)
-where
-    F: FnMut(usize) -> Option<syn::TypeParamBound>,
-{
-    match typ {
-        syn::Type::Path(syn::TypePath { path, .. }) => {
-            if let Some(ident) = path.get_ident() {
-                // The type is a single ident (no other path segments, has no generics):
-                // call back if it is a generic type for us
-                for generic in &mut generics.params {
-                    if let syn::GenericParam::Type(type_param) = generic {
-                        if &type_param.ident == ident {
-                            if let Some(bound) = func(depth) {
-                                type_param.bounds.push(bound);
-                            }
-                        }
-                    }
-                }
-            } else {
-                // Analyze the type parameters of the type, as they may be generics for us as well
-                // This tries to reproduce the bounds that field types place on
-                // their generic types, directly or indirectly. For this the API depth (the const generic
-                // param to `TreeKey<Y>` etc) is determined as follows:
-                //
-                // Assume that all types use their generic T at
-                // relative depth 1, i.e.
-                // * if `#[tree(depth(Y > 1))] a: S<T>` then `T: Tree{Key,Serialize,Deserialize}<Y - 1>`
-                // * else (that is if `Y = 1` or `a: S<T>` without `#[tree]`) then
-                //   `T: serde::{Serialize,Deserialize}`
-                //
-                // And analogously for nested types `S<T<U>>` and `[[T; ..]; ..]` etc.
-                // This is correct for all types in this library (Option, array, structs with the derive macro).
-                //
-                // The bounds are conservative (might not be required) and
-                // fragile (might apply the wrong bound).
-                // This matches the standard derive behavior and its issues
-                // https://github.com/rust-lang/rust/issues/26925
-                //
-                // To fix this, one would extend the attribute syntax to allow overriding bounds.
-                for seg in path.segments.iter() {
-                    if let syn::PathArguments::AngleBracketed(args) = &seg.arguments {
-                        for arg in args.args.iter() {
-                            if let syn::GenericArgument::Type(typ) = arg {
-                                // Found type argument in field type: recurse
-                                walk_type_params(typ, func, depth.saturating_sub(1), generics);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        syn::Type::Array(syn::TypeArray { elem, .. })
-        | syn::Type::Slice(syn::TypeSlice { elem, .. }) => {
-            // An array or slice places the element exactly one level deeper: recurse.
-            walk_type_params(elem, func, depth.saturating_sub(1), generics);
-        }
-        syn::Type::Reference(syn::TypeReference { elem, .. }) => {
-            // A reference is transparent
-            walk_type_params(elem, func, depth, generics);
-        }
-        other => panic!("Unsupported type: {:?}", other),
-    };
 }
